sprinkle ThrowIfCancellationRequested in the code, add some missing tokens to mailKit calls, and optimize some queries
